feat: unify backends with QT
